requirements.txt
django>=5.0,<6.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
obs-websocket-py
